+++ app/pages/dashboard/PatientDrawerContent/index.js
@@ -1,0 +1,3 @@
+import PatientDrawerContent from "./PatientDrawerContent";
+
+export default PatientDrawerContent;
