+++ app/pages/dashboard/PatientDrawerContent/useAGPImages/getOpts.js
