+++ app/pages/dashboard/PatientDrawerContent/useAGPImages/index.js
@@ -1,0 +1,5 @@
+import useAGPImages, { STATUS } from './useAGPImages';
+
+export { STATUS };
+
+export default useAGPImages;
